fix: App.tsxの関数の責務を分離
